refactor(template): push client boundaries to leaves in order components
- order-card.tsx: Remove "use client", fix React namespace import to named
  import, simplify OrderCardItemCount to render children (translations now
  resolved on server in page.tsx)
- order-progress.tsx: Remove "use client", fix React namespace import,
  extract OrderProgressComposed to order-progress-client.tsx
- order-filters.tsx: Extract OrderFiltersComposed to order-filters-client.tsx
